Added appInterface documentation
- Added a new documentation page for appInterface and reorganize the _START_HERE.md file to detail the flow of the program execution and move the To-Do List to the bottom.
- Updated appInterface.h to declare example helper functions and separate the main/necessary variables from the example ones.
- Refactored appInterface.cpp to split update logic into new functions updateFpsCounter() and updateMeshRotation(), slightly improve the initialization error text, and add explanatory comments.
- Avoided warnings in appRuntime_win32.cpp by manually casting window dimensions (unsigned int) to LONG.

+++ ibex3D_sampleApplication/source/ibex3D/core/appRuntime_win32.cpp
@@ -226,7 +226,7 @@ bool appRuntime::initWindow(unsigned int wndWidth, unsigned int wndHeight, const
 
 	DWORD wndStyle = WS_OVERLAPPEDWINDOW;
 
-	RECT wndRect = { 0, 0, wndWidth, wndHeight };
+	RECT wndRect = { 0, 0, static_cast<LONG>(wndWidth), static_cast<LONG>(wndHeight)};
 	AdjustWindowRect(&wndRect, wndStyle, FALSE);
 
 	wndData->hWnd = CreateWindowExA
